Added Radar Plot Generation and fixed DB script off by one issue
Added the ability to generate radar plots of symptom prevalence based on DB patients.  Right now, the plot is created as a JPEG and output in "server\allPatientSymptoms.jpg"  Also fixed an issue where the PATIENT_SYMPTOMS and SYMPTOMS tables were mismatched because SYMPTOM_ID field was off by one in one of the tabled.